+++ io/read.go
@@ -3,30 +3,28 @@ package io
 import (
 	"encoding/json"
 	"os"
-
-	"github.com/zeroibot/pack/ds"
 )
 
 // ReadJSON reads a JSON object of given type from given path
-func ReadJSON[T any](path string) ds.Result[T] {
+func ReadJSON[T any](path string) (T, error) {
+	var item T
 	bytes, err := os.ReadFile(path)
 	if err != nil {
-		return ds.Error[T](err)
+		return item, err
 	}
-	var item T
 	err = json.Unmarshal(bytes, &item)
 	if err != nil {
-		return ds.Error[T](err)
+		return item, err
 	}
-	return ds.NewResult(item, nil)
+	return item, nil
 }
 
 // ReadJSONList reads a JSON list of given type from given path
-func ReadJSONList[T any](path string) ds.Result[[]T] {
+func ReadJSONList[T any](path string) ([]T, error) {
 	return ReadJSON[[]T](path)
 }
 
 // ReadJSONMap reads a JSON map of given value type from given path
-func ReadJSONMap[V any](path string) ds.Result[map[string]V] {
+func ReadJSONMap[V any](path string) (map[string]V, error) {
 	return ReadJSON[map[string]V](path)
 }