+++ model/schema_flag.go
@@ -3,7 +3,6 @@ package model
 import (
 	"database/sql"
 
-	"github.com/zeroibot/pack/ds"
 	"github.com/zeroibot/pack/fail"
 	"github.com/zeroibot/pack/my"
 	"github.com/zeroibot/pack/qb"
@@ -67,19 +66,20 @@ func (s *Schema[T]) setFlagsAt(rq *my.Request, condition qb.DualCondition[T], fi
 	qb.Update(this, q, field, flag)
 
 	// Execute UpdateQuery
-	var result ds.Result[sql.Result]
+	var result sql.Result
+	var err error
 	if isTx {
 		rq.AddTxStep(q)
 		checker := qb.AssertRowsAffected(numItems)
-		result = qb.ExecTx(q, rq.Tx, checker)
+		result, err = qb.ExecTx(q, rq.Tx, checker)
 	} else {
-		result = qb.Exec(q, rq.DB)
+		result, err = qb.Exec(q, rq.DB)
 	}
-	if result.IsError() {
+	if err != nil {
 		rq.Fail(my.Err500, "Failed to update %s flag", s.Name)
-		return result.Error()
+		return err
 	}
-	rowsUpdated := qb.RowsAffected(result.Value())
+	rowsUpdated := qb.RowsAffected(result)
 
 	// If not transaction, check if rowsUpdated == numItems
 	if !isTx && rowsUpdated != numItems {